test(consensus): replace manually serialized test-vectors with dynamically generated ones
Previously there were hard-coded serialized versions of
blocks/transactions. This is problematic, since the block header or the
tx format may change, and this would require manual regeneration of the
test data each time. Instead, we add a helper function that tries to
load them from a file, or calculates them and writes them to a file if
they do not already exist. If the header/tx format is changed, the file
must be deleted, so that the function regenerates the most up to date
version of it