Update decrypt.rs
possible error in the 'parse()' function. Specifically, the format!() call in the last line returns a formatted string, but the string is not returned from the function. Instead, it should be returned like this:

Ok(format!("✅ Created deployment transaction for '{}'", self.program_id.to_string().bold()))
Without the Ok(), the formatted string will not be returned as a Result<String>, which is the expected return type of the function.

Signed-off-by: liz atkinson <126437534+lizatkinson@users.noreply.github.com>