Tighten iput validation on ping messages.
When a ping message is serialized, the absence or presence of the optional block
locators structure is encoded as either a 0 or a 1 byte.

In deserialization, any non-zero byte is treated the same as a 1 byte for the
purpose of indicating absence or presence of the optional block locators
structure.

This is not necessarily a bug, if the (implicit) specification of the format of
a ping message is that that byte may have any value, and that 0 indicates
absence of block locators while non-0 indicates their presence. However, one
might argue that messages should be represented in unique ways, so that
serialization and deserialization are always each other's inverses.

If in the future the format of a ping message is extended so that that byte can
have other values besides 0 or 1 (e.g. 2), then the deserialization could
silently (attempt to) interpret byte 2 as if it were byte 1.

Unique message representations and tight input validation are in line with the
LangSec approach.