Update README.md
Since the build guide does not mention an ability to use snarkOS flags with `.sh`-files (specifically --network), it has been added to the guide with appropriate network values.

Also, if anyone wants to NOT rebuild 'snarkOS' binary at startup, I've added a straightforward description with `snarkos`. A disclaimer is also added if anyone decides to run the node this way.